:recycle: Split up referral token schema
